rag_backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
langchain
langchain-openai
chromadb
sentence-transformers
rank-bm25
pypdf
docx2txt
pandas
pillow
pytesseract
openpyxl
python-magic